Consolidate auth modules
