+++ db/README.md
@@ -32,7 +32,7 @@ exists. PostgreSQL enforces those constraints for new writes immediately without
 rejecting the deployment because of an unknown historical orphan.
 
 Migration `0001` adopts the singular `user`, `session`, `account`, and
-`verification` tables previously managed from `auth.ts`. It preserves the
+`verification` tables previously managed from `auth/index.ts`. It preserves the
 existing `timestamptz` representation and rows, safely adds the nullable
 phone-number plugin fields when absent, and installs the indexes used by Better
 Auth. The runtime now assumes versioned migrations have run before requests are
