+++ tests/database-schema.test.ts
@@ -169,7 +169,7 @@ describe("migration deployment policy", () => {
       )
     );
     const authSource = await readFile(
-      new URL("../auth.ts", import.meta.url),
+      new URL("../auth/index.ts", import.meta.url),
       "utf8"
     );
     const authMigration = await readFile(
